Refactoring names in client_side_validation
