+++ client_side_validation/src/lib.rs
@@ -42,12 +42,12 @@ extern crate amplify_derive;
 extern crate bitcoin_hashes;
 
 #[macro_use]
-mod client_side_validation;
+mod commit_encode;
 pub mod commit_verify;
 mod digests;
 pub mod single_use_seals;
 
-pub use crate::client_side_validation::{
-    commit_strategy, merklize, CommitEncode, CommitEncodeWithStrategy, Conceal,
-    ConsensusCommit, MerkleNode,
+pub use crate::commit_encode::{
+    commit_strategy, merklize, CommitConceal, CommitEncode,
+    CommitEncodeWithStrategy, ConsensusCommit, MerkleNode,
 };
